Initial Commit
Hi I'm Jonathan Renan Esponilla and you're watching Template.
This is an HTML and CSS template that I made for the purposes of
quickly starting up web projects if I want to.

If you're looking back at this commit and seeing things
related to "Starry Pop Dreams" that's because
that's the name of the project this template was made for.
Please ignore that I'll remove them later.